+++ surveys/views_api.py
@@ -18,6 +18,8 @@ class SurveyView(generics.RetrieveAPIView):
 
 
 class SurveyCreateMixin:
+    serializer_class = SurveyResponseSerializer
+
     def dispatch(self, request, *args, **kwargs):
         survey_pk = kwargs.get('pk')
         self.user_id = kwargs.get('user_id')
@@ -29,14 +31,13 @@ class SurveyCreateMixin:
 
 
 class SurveyCreateView(SurveyCreateMixin, generics.CreateAPIView):
-    serializer_class = SurveyResponseSerializer
+    pass
 
 
 class SurveyGetLatestCreateView(
         SurveyCreateMixin,
         generics.RetrieveAPIView,
         generics.CreateAPIView):
-    serializer_class = SurveyResponseSerializer
     queryset = UserResponse.objects.all()
 
     def get(self, request, *args, **kwargs):