+++ web/app/scripts/compiled-es5/libraries.js
@@ -1,0 +1,6 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+var angular = exports.angular = window.angular;