+++ web/app/scripts/compiled-es5/services/api.js
@@ -1,0 +1,60 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+    value: true
+});
+exports.moduleProperties = undefined;
+
+var _libraries = require('./../libraries.js');
+
+var _projectConfig = require('./../providers/project-config.js');
+
+var _projectConfig2 = _interopRequireDefault(_projectConfig);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+var moduleProperties = exports.moduleProperties = {
+    moduleName: 'incuna-surveys.api',
+    componentName: 'API'
+};
+
+var _module = _libraries.angular.module(moduleProperties.moduleName, [_projectConfig2.default.moduleName]);
+
+_module.service(moduleProperties.componentName, ['$http', _projectConfig2.default.componentName, function ($http, ProjectConfig) {
+    return {
+        getBaseUrl: function getBaseUrl() {
+            var endpoint = 'forms';
+            var apiRoot = ProjectConfig.apiRoot;
+            return apiRoot + '/' + endpoint;
+        },
+        getList: function getList() {
+            var url = this.getBaseUrl();
+
+            return $http.get(url).then(function (response) {
+                return response.data;
+            });
+        },
+        getForm: function getForm(url) {
+            return $http.get(url).then(function (response) {
+                return response.data;
+            });
+        },
+        get: function get(url) {
+            return $http.get(url).then(function (response) {
+                return response.data;
+            });
+        },
+        post: function post(url, responses) {
+            var data = {
+                // jscs:disable disallowQuotedKeysInObjects
+                'user_responses': responses
+                // jscs:enable disallowQuotedKeysInObjects
+            };
+            return $http.post(url, data).then(function (response) {
+                return response.data;
+            });
+        }
+    };
+}]);
+
+exports.default = moduleProperties;